# specs/001-phi-redaction-proxy/contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: phi-redactor Proxy API
  version: 0.1.0
  description: |
    PHI Redaction Proxy for AI/LLM Interactions.
    Provides transparent PHI redaction between client applications and LLM APIs.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local proxy server
paths:
...
  /openai/v1/chat/completions:
    post:
      operationId: openaiChatCompletions
      summary: Proxy OpenAI chat completions with PHI redaction
      description: |
        Intercepts chat completion requests, redacts PHI from messages,
        forwards to OpenAI, and re-hydrates PHI in the response.
      tags: [OpenAI Proxy]
      parameters:
        - $ref: "#/components/parameters/XSessionId"
        - $ref: "#/components/parameters/XApiKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenAIChatRequest"
      responses:
        "200":
          description: Chat completion response (with PHI re-hydrated)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenAIChatResponse"
            text/event-stream:
              schema:
                type: string
                description: SSE stream of chat completion chunks (with PHI re-hydrated)
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/UpstreamError"
  /openai/v1/embeddings:
    post:
      operationId: openaiEmbeddings
      summary: Proxy OpenAI embeddings with PHI redaction
      tags: [OpenAI Proxy]
      parameters:
        - $ref: "#/components/parameters/XSessionId"
        - $ref: "#/components/parameters/XApiKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenAIEmbeddingsRequest"
      responses:
        "200":
          description: Embeddings response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenAIEmbeddingsResponse"

  # ============================================================
  # Anthropic-Compatible Proxy Endpoints
  # ============================================================
...
  /v1/chat/completions:
    post:
      operationId: defaultChatCompletions
      summary: Default provider chat completions proxy
      description: Routes to the configured default provider.
      tags: [Default Proxy]
      parameters:
        - $ref: "#/components/parameters/XSessionId"
        - $ref: "#/components/parameters/XApiKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenAIChatRequest"
      responses:
        "200":
          description: Chat completion response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenAIChatResponse"
            text/event-stream:
              schema:
                type: string

  # ============================================================
  # Management API
  # ============================================================
...
components:
  parameters:
    XSessionId:
      name: X-Session-Id
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: |
        Session identifier for multi-turn conversation tracking.
        If omitted, a new session is created automatically.
    XApiKey:
      name: Authorization
      in: header
      required: true
      schema:
        type: string
      description: LLM provider API key (forwarded to upstream)
  schemas:
    OpenAIChatRequest:
      type: object
      required: [model, messages]
      properties:
        model:
          type: string
        messages:
          type: array
          items:
            type: object
            required: [role, content]
            properties:
              role:
                type: string
                enum: [system, user, assistant]
              content:
                type: string
        stream:
          type: boolean
          default: false
        temperature:
          type: number
        max_tokens:
          type: integer
    OpenAIChatResponse:
      type: object
      properties:
        id:
          type: string
        object:
          type: string
          enum: [chat.completion]
        choices:
          type: array
          items:
            type: object
            properties:
              index:
                type: integer
              message:
                type: object
                properties:
                  role:
                    type: string
                  content:
                    type: string
              finish_reason:
                type: string
        usage:
          type: object
          properties:
            prompt_tokens:
              type: integer
            completion_tokens:
              type: integer
            total_tokens:
              type: integer
        x_phi_redactor:
          $ref: "#/components/schemas/RedactionMetadata"
    OpenAIEmbeddingsRequest:
      type: object
      required: [model, input]
      properties:
        model:
          type: string
        input:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
    OpenAIEmbeddingsResponse:
      type: object
      properties:
        object:
          type: string
        data:
          type: array
          items:
            type: object
            properties:
              object:
                type: string
              index:
                type: integer
              embedding:
                type: array
                items:
                  type: number
...
    RedactionMetadata:
      type: object
      description: PHI redaction metadata appended to proxied responses
      properties:
        session_id:
          type: string
          format: uuid
        redactions_in_request:
          type: integer
        rehydrations_in_response:
          type: integer
        processing_time_ms:
          type: number
        categories_detected:
          type: array
          items:
            type: string
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              detail:
                type: string
    UpstreamError:
      description: Upstream LLM API error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              upstream_status:
                type: integer
              detail:
                type: string